Feat: mv /[locale], updated routing to take the locale into account
