+++ src/app/[locale]/favicon.ico
