+++ src/components/admin/sellerProducts/SellerTabList.tsx
@@ -1,7 +1,8 @@
 import { SellerComplete, Status } from "@/api";
 import { TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { useRouter } from "@/i18n/navigation";
 import { useTranslations } from "next-intl";
-import { useRouter, useSearchParams } from "next/navigation";
+import { useSearchParams } from "next/navigation";
 
 interface SellerTabListProps {
   status: Status;
