+++ src/components/user/PaymentButton.tsx
@@ -1,6 +1,6 @@
 import { usePaymentUrl } from "@/hooks/usePaymentUrl";
+import { useRouter } from "@/i18n/navigation";
 import { useTranslations } from "next-intl";
-import { useRouter } from "next/navigation";
 import { useState } from "react";
 
 import { HelloAssoButton } from "../custom/HelloAssoButton";
